backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Agent Framework
letta

# Vector Database
chromadb==0.5.11

# LLM
anthropic==0.37.1

# Web scraping
beautifulsoup4==4.12.3
playwright==1.48.0
lxml==5.3.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
